Add ability to add handlers for new resources to AdmissionController via RegisterHandler().
This PR refactors AdmissionController and decouples its core logic from
logic responsible for handling resources. Handlers for already
supported resources are registered via the new mechanism.
